мердж двух систем стола с картой
